introduce ExecStmtRuntime and copy the implementation
